+++ package.json
@@ -9,7 +9,6 @@
     "@testing-library/user-event": "^12.1.10",
     "axios": "^0.21.0",
     "bootstrap": "^4.5.3",
-    "bs-stepper": "^1.7.0",
     "dotenv": "^8.2.0",
     "jquery": "^3.5.1",
     "react": "^17.0.1",
