+++ src/App.jsx
@@ -5,16 +5,17 @@ import {
   Switch
 } from 'react-router-dom'
 import 'bootstrap/dist/css/bootstrap.min.css'
-import SignInPage from './pages/signInPage'
-import { SignUpPage } from './pages/signUpPage'
-import { MainPage } from './pages/MainPage/MainPage'
-import { RegisterPage } from './pages/registerPage'
-import { ConfirmedPage } from './pages/confirmedPage'
+import SignInPage from './pages/signIn/SignInPage'
+import { SignUpPage } from './pages/signUp/SignUpPage'
+import { MainPage } from './pages/main/MainPage'
+import { RegisterPage } from './pages/register/RegisterPage'
+import { ConfirmedPage } from './pages/confirmed/ConfirmedPage'
+import DashboardPage from './pages/dashboard/DashBoardPage'
+import { RegistryPage } from './pages/student/RegistryPage'
 import path from 'path'
 import dotenv from 'dotenv'
 
 import { UserContextProvider } from './context/UserContext'
-import DashboardPage from './pages/dashboardPage/DashBoardPage'
 import { StudentContextProvider } from './context/StudentContext'
 import { CompanyContextProvider } from './context/CompanyContext'
 
@@ -28,7 +29,6 @@ function App () {
     <UserContextProvider>
       <StudentContextProvider>
         <CompanyContextProvider>
-
           <Router>
             <Switch>
               <Redirect exact from="/" to="/main" />
@@ -39,6 +39,7 @@ function App () {
                 <SignUpPage />
               </Route>
               <Route path="/register/:accountType" component={RegisterPage} />
+              <Route path="/student/register/:section" component={RegistryPage} />
               <Route path="/dashboard">
                 <DashboardPage />
               </Route>
