--- src/components/Detail/Button/styles.css
@@ -1,8 +1,0 @@
-.title-style {
-  color: #47525E;
-  font-family: Lato;
-  font-size: 40px;
-  line-height: 51px;
-  width: 100%;
-  text-align: center;
-}
