+++ src/components/Register/Student/QualificationRegister/QualificationRegister.jsx
