+++ src/pages/signUp/SignUpPage.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { SignUp } from '../../components/SignUp'
+import { SignUp } from '../../components/SignUp/SignUp'
 import './signUp.css'
 
 export const SignUpPage = () => {
